Refactors and documents project for improved maintainability
Restructures the project with a clear folder organization and updated file paths.

Adds comprehensive documentation, including a detailed README, API reference, and data format specifications.

Enhances the `.gitignore` file and implements a changelog for tracking project versions.
