+++ CHANGELOG.md
@@ -1,0 +1,74 @@
+# Changelog
+
+All notable changes to the DHL Staff Dashboard project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [Unreleased]
+
+### Added
+
+- Comprehensive project restructuring
+- Detailed README.md with full documentation
+- Proper folder organization (src/, docs/, data/, etc.)
+- Enhanced .gitignore file
+- Project changelog
+
+### Changed
+
+- Moved JavaScript files to `src/` directory
+- Organized UI components in `src/components/`
+- Organized utility functions in `src/utils/`
+- Updated all import paths to reflect new structure
+- Moved CSV data files to `data/` directory
+- Moved documentation media to `docs/` directory
+
+### Fixed
+
+- Import path references in all JavaScript modules
+- File organization for better maintainability
+
+## [1.0.0] - Previous Version
+
+### Features
+
+- Employee productivity dashboard
+- Performance chart visualization
+- UPH (Units Per Hour) tracking
+- Interactive data table
+- Real-time data filtering
+- Break time configuration
+- Staff performance analytics
+- Bootstrap-based responsive design
+- CSV data processing
+- Chart.js integration
+- Toast notifications
+- Date range filtering
+
+### Components
+
+- Main dashboard interface
+- Performance and UPH charts
+- Data table with sorting
+- Staff configuration system
+- CSV data parser
+- Filter controls
+- Loading indicators
+
+---
+
+## Release Notes
+
+### Version History
+
+- **v1.0.0**: Initial dashboard implementation
+- **v1.1.0**: Project restructuring and documentation
+
+### Upcoming Features
+
+- Real-time data updates
+- Enhanced mobile experience
+- Additional chart types
+- User authentication
+- Advanced analytics
