+++ src/components/table.js
@@ -1,14 +1,14 @@
-import { capitalize, toHoursAndMinutes } from "../scripts/helpers.js";
+import { capitalize, toHoursAndMinutes } from "../utils/helpers.js";
 import { tableCheckboxElem, tableDiv } from "./elements.js";
 
 export const makeTable = (packers) => {
   if (!tableCheckboxElem.checked) {
     tableDiv.className = "display-none";
   }
-  packers.sort((a, b) => b.numberOfItems.length - a.numberOfItems.length)
-  const tableElement = document.querySelector('#table');
-  const tableHtml =
-    [`
+  packers.sort((a, b) => b.numberOfItems.length - a.numberOfItems.length);
+  const tableElement = document.querySelector("#table");
+  const tableHtml = [
+    `
   <div class="table-responsive p-2 text-center shadow-sm">
 		<table class="table-sm align-middle table table-hover">
 			<thead>
@@ -21,26 +21,35 @@ export const makeTable = (packers) => {
 					<th scop="col">UPH</th>
 					<th scop="col">Total Time</th>
           </tr>
-          </thead><tbody>`];
+          </thead><tbody>`,
+  ];
 
-  const grandTotalHr = packers.reduce((a, v) => a + v['totalTimeInHr'], 0);
-  const grandTotalUph = packers.reduce((a, v) => a + v['uph'], 0);
-  const grandTotalUpo = packers.reduce((a, v) => a + v['upo'], 0);
-  const grandTotalItems = packers.reduce((a, v) => a + v['numberOfItems'].length, 0);
-  const grandTotalOrders = packers.reduce((a, v) => a + Object.values(v['numberOfOrders']).length, 0);
+  const grandTotalHr = packers.reduce((a, v) => a + v["totalTimeInHr"], 0);
+  const grandTotalUph = packers.reduce((a, v) => a + v["uph"], 0);
+  const grandTotalUpo = packers.reduce((a, v) => a + v["upo"], 0);
+  const grandTotalItems = packers.reduce(
+    (a, v) => a + v["numberOfItems"].length,
+    0
+  );
+  const grandTotalOrders = packers.reduce(
+    (a, v) => a + Object.values(v["numberOfOrders"]).length,
+    0
+  );
 
-  tableHtml.push(packers.map(packer => {
-    const {
-      packerName,
-      packerId,
-      numberOfItems,
-      numberOfOrders,
-      totalTimeInHr,
-      upo,
-      uph
-    } = packer
+  tableHtml.push(
+    packers
+      .map((packer) => {
+        const {
+          packerName,
+          packerId,
+          numberOfItems,
+          numberOfOrders,
+          totalTimeInHr,
+          upo,
+          uph,
+        } = packer;
 
-    return `
+        return `
     <tr>
     <th scop="row">${capitalize(packerName)}</th>
     <td>${packerId}</td>
@@ -49,9 +58,10 @@ export const makeTable = (packers) => {
     			<td>${upo.toFixed(2)}</td>
     			<td>${uph.toFixed(2)}</td>
     			<td>${toHoursAndMinutes(totalTimeInHr)} hr</td>
-    	</tr>`
-  }).join(""))
-
+    	</tr>`;
+      })
+      .join("")
+  );
 
   tableHtml.push(`</tbody><tfoot>
   			<tr class="table-success">
@@ -62,7 +72,7 @@ export const makeTable = (packers) => {
   					<th>-</th>
   					<th>-</th>
   					<th>${toHoursAndMinutes(grandTotalHr)} hr</th>
-  			</tr>`)
+  			</tr>`);
   tableHtml.push(`
   			<tr class="table-warning">
         <th>Grand Average</th>
@@ -71,10 +81,10 @@ export const makeTable = (packers) => {
         <th>${(grandTotalOrders / packers.length).toFixed(2)}</th>
         <th>${(grandTotalUpo / packers.length).toFixed(2)}</th>
         <th>${(grandTotalUph / packers.length).toFixed(2)}</th>
-        <th>${toHoursAndMinutes(Math.floor(grandTotalHr / packers.length))} hr</th>
-  			</tr>`)
-  tableHtml.push('</tfoot></table></div>')
-  tableElement.innerHTML = tableHtml.join('');
-
-}
-
+        <th>${toHoursAndMinutes(
+          Math.floor(grandTotalHr / packers.length)
+        )} hr</th>
+  			</tr>`);
+  tableHtml.push("</tfoot></table></div>");
+  tableElement.innerHTML = tableHtml.join("");
+};
